refactering the code 2.2 removed persistance folder,moved  enity and dao folder from persiatnce to root directory

+++ src/main/java/com/sabtok/services/BookService.java
@@ -4,14 +4,14 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import com.sabtok.persistance.entity.Book;
+import com.sabtok.entity.Book;
 
 @Service
 public interface BookService {
 
 	public Book saveBook(Book book);
 	public String updateBook(Book book);
-	public Integer bookNumber();
+	public Long bookNumber();
 	public List<Book> getBooks();
 	public Book getBookDetails(String bookId);
